change constants from the game component, moved to other folders
