+++ src/components/Game/option.constants.js
@@ -1,0 +1,18 @@
+import deleteOption from '../../assets/options/delete.png'
+import rotateOption from '../../assets/options/rotate.png'
+import moveOption from '../../assets/options/move.png'
+
+export const editionOptions = [
+  {
+    image: deleteOption,
+    title: 'Delete',
+  },
+  {
+    image: rotateOption,
+    title: 'Rotate',
+  },
+  {
+    image: moveOption,
+    title: 'Move',
+  },
+]
